+++ lib/slim.dart
@@ -3,5 +3,5 @@ library slim;
 export 'src/state_management.dart';
 export 'src/extensions.dart';
 export 'src/localization.dart';
-export 'src/rest_api.dart';
+export 'src/api.dart';
 export 'src/builders.dart';